Remove base abstract class
