+++ src/functions/definitions/connector-function.ts
@@ -1,26 +1,46 @@
-import {
-  ManifestFunctionSchema,
-  ManifestFunctionType,
-} from "../../manifest/manifest_schema.ts";
+import { ManifestFunctionType } from "../../manifest/manifest_schema.ts";
 import {
   ParameterSetDefinition,
   PossibleParameterKeys,
 } from "../../parameters/types.ts";
-import { FunctionDefinitionArgs } from "../types.ts";
-import { FunctionDefinition } from "./function.ts";
+import { FunctionDefinitionArgs, ISlackFunctionDefinition } from "../types.ts";
+
+/**
+ * Define a connector and its input and output parameters for use in a Slack application.
+ * @param {FunctionDefinitionArgs<InputParameters, OutputParameters, RequiredInput, RequiredOutput>} definition Defines information about your function (title, description) as well as formalizes the input and output parameters of a connector
+ * @returns {ConnectorFunctionDefinition}
+ */
+export const DefineConnector = <
+  InputParameters extends ParameterSetDefinition,
+  OutputParameters extends ParameterSetDefinition,
+  RequiredInput extends PossibleParameterKeys<InputParameters>,
+  RequiredOutput extends PossibleParameterKeys<OutputParameters>,
+>(
+  definition: FunctionDefinitionArgs<
+    InputParameters,
+    OutputParameters,
+    RequiredInput,
+    RequiredOutput
+  >,
+) => {
+  return new ConnectorFunctionDefinition(definition);
+};
 
 export class ConnectorFunctionDefinition<
   InputParameters extends ParameterSetDefinition,
   OutputParameters extends ParameterSetDefinition,
   RequiredInput extends PossibleParameterKeys<InputParameters>,
   RequiredOutput extends PossibleParameterKeys<OutputParameters>,
-> extends FunctionDefinition<
-  InputParameters,
-  OutputParameters,
-  RequiredInput,
-  RequiredOutput
-> {
+> implements
+  ISlackFunctionDefinition<
+    InputParameters,
+    OutputParameters,
+    RequiredInput,
+    RequiredOutput
+  > {
   type: ManifestFunctionType = "API";
+  id: string;
+
   constructor(
     public definition: FunctionDefinitionArgs<
       InputParameters,
@@ -29,18 +49,7 @@ export class ConnectorFunctionDefinition<
       RequiredOutput
     >,
   ) {
-    super(definition);
-  }
-
-  export(): ManifestFunctionSchema {
-    return {
-      title: this.definition.title,
-      description: this.definition.description,
-      type: this.type,
-      input_parameters: this.definition.input_parameters ??
-        { properties: {}, required: [] },
-      output_parameters: this.definition.output_parameters ??
-        { properties: {}, required: [] },
-    };
+    this.id = definition.callback_id;
+    this.definition = definition;
   }
 }
