+++ src/functions/definitions/slack-function.ts
@@ -2,25 +2,52 @@ import {
   ManifestFunctionSchema,
   ManifestFunctionType,
 } from "../../manifest/manifest_schema.ts";
+import { SlackManifest } from "../../mod.ts";
 import {
   ParameterSetDefinition,
   PossibleParameterKeys,
 } from "../../parameters/types.ts";
-import { SlackFunctionDefinitionArgs } from "../types.ts";
-import { FunctionDefinition } from "./function.ts";
+import {
+  ISlackFunctionDefinition,
+  SlackFunctionDefinitionArgs,
+} from "../types.ts";
+
+/**
+ * Define a function and its input and output parameters for use in a Slack application.
+ * @param {SlackFunctionDefinitionArgs<InputParameters, OutputParameters, RequiredInput, RequiredOutput>} definition Defines information about your function (title, description) as well as formalizes the input and output parameters of your function
+ * @returns {SlackFunctionDefinition}
+ */
+export const DefineFunction = <
+  InputParameters extends ParameterSetDefinition,
+  OutputParameters extends ParameterSetDefinition,
+  RequiredInput extends PossibleParameterKeys<InputParameters>,
+  RequiredOutput extends PossibleParameterKeys<OutputParameters>,
+>(
+  definition: SlackFunctionDefinitionArgs<
+    InputParameters,
+    OutputParameters,
+    RequiredInput,
+    RequiredOutput
+  >,
+) => {
+  return new SlackFunctionDefinition(definition);
+};
 
 export class SlackFunctionDefinition<
   InputParameters extends ParameterSetDefinition,
   OutputParameters extends ParameterSetDefinition,
   RequiredInput extends PossibleParameterKeys<InputParameters>,
   RequiredOutput extends PossibleParameterKeys<OutputParameters>,
-> extends FunctionDefinition<
-  InputParameters,
-  OutputParameters,
-  RequiredInput,
-  RequiredOutput
-> {
+> implements
+  ISlackFunctionDefinition<
+    InputParameters,
+    OutputParameters,
+    RequiredInput,
+    RequiredOutput
+  > {
   type: ManifestFunctionType = "app";
+  id: string;
+
   constructor(
     public definition: SlackFunctionDefinitionArgs<
       InputParameters,
@@ -29,7 +56,15 @@ export class SlackFunctionDefinition<
       RequiredOutput
     >,
   ) {
-    super(definition);
+    this.id = definition.callback_id;
+    this.definition = definition;
+  }
+
+  registerParameterTypes(manifest: SlackManifest) {
+    const { input_parameters: inputParams, output_parameters: outputParams } =
+      this.definition;
+    manifest.registerTypes(inputParams?.properties ?? {});
+    manifest.registerTypes(outputParams?.properties ?? {});
   }
 
   export(): ManifestFunctionSchema {
@@ -43,4 +78,32 @@ export class SlackFunctionDefinition<
         { properties: {}, required: [] },
     };
   }
+}
+
+export function isCustomFunction<
+  InputParameters extends ParameterSetDefinition,
+  OutputParameters extends ParameterSetDefinition,
+  RequiredInput extends PossibleParameterKeys<InputParameters>,
+  RequiredOutputs extends PossibleParameterKeys<OutputParameters>,
+>(
+  functionDefinition: ISlackFunctionDefinition<
+    InputParameters,
+    OutputParameters,
+    RequiredInput,
+    RequiredOutputs
+  >,
+): functionDefinition is SlackFunctionDefinition<
+  InputParameters,
+  OutputParameters,
+  RequiredInput,
+  RequiredOutputs
+> {
+  if (
+    functionDefinition.type === "app" &&
+    functionDefinition.export &&
+    functionDefinition.registerParameterTypes
+  ) {
+    return true;
+  }
+  return false;
 }
